requirements.txt
crispy-bootstrap5==0.6
Django==4.0.4
django-countries==7.3.2
django-crispy-forms==1.14.0
django-extensions==3.1.5
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
